Rename `internal/protocol` to `internal/impl`.
