+++ internal/impl/default.go
@@ -1,4 +1,4 @@
-package protocol
+package impl
 
 import (
 	"net"
