+++ internal/impl/header.go
@@ -1,4 +1,4 @@
-package protocol
+package impl
 
 const (
 	headerUserAgent = "User-Agent"
