+++ internal/impl/payload.go
@@ -1,4 +1,4 @@
-package protocol
+package impl
 
 import (
 	"bytes"
